+++ packages/web/src/lib/web-mode.ts
@@ -1,4 +1,4 @@
-import algosdk, { SignedTransaction, SuggestedParams, Transaction } from "algosdk";
+import algosdk, { SignedTransaction, EncodedSignedTransaction, SuggestedParams, Transaction } from "algosdk";
 
 import { AlgoSigner, JsonPayload, WalletTransaction } from "../algo-signer-types";
 import { BuilderError, ERRORS } from "../errors/errors";
@@ -147,6 +147,61 @@ export class WebMode {
 				: Number(userParams.firstValid) + Number(userParams.validRounds);
 
 		return s;
+	}
+
+	/**
+	 * Appends signature to a multisig transaction using algosigner
+	 * @param txns array of multisignature transaction with at least one signature
+	 * @param signers a subset of addresses to sign the transaction
+	 * return an object containing a blob attribute encoded in base64
+	 */
+	async appendSignMultisigTransaction(txns: WalletTransaction[], signers: string[]): Promise<JsonPayload> {
+		const result:JsonPayload = {};
+		for (let i = 0; i < txns.length; ++i) {
+			const txn = txns[i];
+			const partialTxn = algosdk.decodeObj(
+				this.algoSigner.encoding.base64ToMsgpack(txn.txn)) as EncodedSignedTransaction; 
+			if (partialTxn.txn === undefined || partialTxn.msig === undefined) {
+				throw new Error("Input transaction must be multisigature transaction signed with at least 1 signature");
+			} 
+			const txnToBeSign = algosdk.Transaction.from_obj_for_encoding(partialTxn.txn);
+			const txnToBeSign_Uint8Array = algosdk.encodeObj(txnToBeSign.get_obj_for_encoding());
+			const txnToBeSign_Base64 = this.algoSigner.encoding.msgpackToBase64(txnToBeSign_Uint8Array);
+
+			const mparams = partialTxn.msig as algosdk.EncodedMultisig;
+			const addrs = mparams.subsig.map((signData) => {
+				return algosdk.encodeAddress(signData.pk);
+			});
+
+			const multisigParams = {
+				version: mparams.v,
+				threshold: mparams.thr,
+				addrs: addrs,
+			}
+
+			const signedTxn = await this.signTransaction([
+				{
+					txn: txnToBeSign_Base64,
+					msig: multisigParams,
+					signers: signers,
+				},
+			]);
+
+			const signedTxnJson = signedTxn[0] as JsonPayload;
+			const blob = signedTxnJson.blob as string;
+
+			const blob1 = this.algoSigner.encoding.base64ToMsgpack(txn.txn);
+			const blob2 = this.algoSigner.encoding.base64ToMsgpack(blob);
+			const combineBlob = algosdk.mergeMultisigTransactions([
+				blob1,
+				blob2,
+			]);
+			const outputBase64 = this.algoSigner.encoding.msgpackToBase64(combineBlob);
+			result[i] = {
+				blob: outputBase64
+			};
+		}
+		return result;
 	}
 
 	/**