+++ CHANGELOG.md
@@ -33,6 +33,10 @@ from `algosdk` and sends it to the network.
 - Added support for `SignedTransaction` object in `executeTx` method in `Runtime`.
 - Added verification for secret key signatures in `Runtime`.
 - Added replace2 and replace3 opcode to `runtime`.
+- Added sha3_256 opcode to `Runtime`
+
+#### @algo-builder/web
+- Added `appendSignMultisigTransaction` function to `WebMode` for appending signature to multisig transaction in the algosigner.
 
 ### Bug Fixes